video

+++ src/main/java/com/video/app/services/impl/CategoryServiceImpl.java
@@ -2,11 +2,12 @@ package com.video.app.services.impl;
 
 import com.video.app.aws.AwsS3Service;
 import com.video.app.entities.Category;
+import com.video.app.exceptions.NotFoundEntity;
 import com.video.app.exceptions.ServiceException;
 import com.video.app.repositories.CategoryRepository;
 import com.video.app.services.CategoryService;
 import com.video.app.utils.DataResponse;
-import com.video.app.utils.ValidString;
+import com.video.app.utils.SlugUtils;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
@@ -35,7 +36,7 @@ public class CategoryServiceImpl implements CategoryService {
         String url = this.awsS3Service.upload(file, "category_image");
         return this.categoryRepository.save(Category.builder()
                 .name(name)
-                .slug(ValidString.slugify(name))
+                .slug(SlugUtils.slugify(name))
                 .deleted(false)
                 .image(url)
                 .build());
@@ -49,25 +50,25 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public DataResponse updateImage(Long id, MultipartFile file) {
         Category category = this.categoryRepository.findById(id)
-                .orElseThrow(() -> new ServiceException("Category not found"));
+                .orElseThrow(() -> new NotFoundEntity("Category not found"));
         String imgUrl = category.getImage();
         String url = this.awsS3Service.upload(file, "category_image");
         category.setImage(url);
-        executorService.submit(() -> {
-            if (!imgUrl.isBlank()) {
+        if (imgUrl != null && !imgUrl.isBlank()) {
+            executorService.submit(() -> {
                 String[] imgUrlSplit = imgUrl.split("/");
                 this.awsS3Service.delete(imgUrlSplit[imgUrlSplit.length - 1], FOLDER_IMAGE_AWS);
-            }
-        });
+            });
+        }
         return new DataResponse("Updated!", category.getImage(), true);
     }
 
     @Override
     public DataResponse updateName(Long id, String name) {
         Category category = this.categoryRepository.findById(id)
-                .orElseThrow(() -> new ServiceException("Category not found"));
+                .orElseThrow(() -> new NotFoundEntity("Category not found"));
         category.setName(name);
-        category.setSlug(ValidString.slugify(name));
+        category.setSlug(SlugUtils.slugify(name));
         this.entityManager.merge(category);
         return new DataResponse("Updated!", null, true);
     }
@@ -75,9 +76,14 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public DataResponse delete(Long id) {
         Category category = this.categoryRepository.findByIdAndDeleted(id, false)
-                .orElseThrow(() -> new ServiceException("Category not found"));
+                .orElseThrow(() -> new NotFoundEntity("Category not found"));
         category.setDeleted(true);
         this.entityManager.merge(category);
         return new DataResponse("Deleted!", null, true);
+    }
+
+    @Override
+    public Category findById(Long id) {
+        return this.categoryRepository.findById(id).orElseThrow(() -> new NotFoundEntity("Category not found"));
     }
 }
